+++ Neo4j.Driver/Neo4j.Driver.Tests/Mapping/BuiltMapperTests.cs
@@ -15,7 +15,7 @@
 
 using System;
 using FluentAssertions;
-using Neo4j.Driver.Preview.Mapping;
+using Neo4j.Driver.Mapping;
 using Neo4j.Driver.Tests.TestUtil;
 using Xunit;
