+++ Neo4j.Driver/Neo4j.Driver.Tests/Mapping/MappableValueProviderTests.cs
@@ -18,7 +18,7 @@ using FluentAssertions;
 using Moq;
 using Moq.AutoMock;
 using Neo4j.Driver.Internal.Types;
-using Neo4j.Driver.Preview.Mapping;
+using Neo4j.Driver.Mapping;
 using Xunit;
 
 namespace Neo4j.Driver.Tests.Mapping;
