+++ Neo4j.Driver/Neo4j.Driver.Tests/Mapping/MappedListCreatorTests.cs
@@ -20,7 +20,7 @@ using System.Linq;
 using FluentAssertions;
 using Moq;
 using Moq.AutoMock;
-using Neo4j.Driver.Preview.Mapping;
+using Neo4j.Driver.Mapping;
 using Neo4j.Driver.Tests.TestUtil;
 using Xunit;
