+++ Neo4j.Driver/Neo4j.Driver.Tests/Mapping/MappingBuilderTests.cs
@@ -15,7 +15,7 @@
 
 using System;
 using FluentAssertions;
-using Neo4j.Driver.Preview.Mapping;
+using Neo4j.Driver.Mapping;
 using Xunit;
 
 namespace Neo4j.Driver.Tests.Mapping;
