+++ Neo4j.Driver/Neo4j.Driver/Public/Mapping/BuiltMapper.cs
@@ -20,7 +20,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Reflection;
 
-namespace Neo4j.Driver.Preview.Mapping;
+namespace Neo4j.Driver.Mapping;
 
 internal class BuiltMapper<T> : IRecordMapper<T>
 {
