+++ Neo4j.Driver/Neo4j.Driver/Public/Mapping/DefaultMapper.cs
@@ -15,10 +15,9 @@
 
 using System;
 using System.Collections.Generic;
-using System.Linq;
 using System.Reflection;
 
-namespace Neo4j.Driver.Preview.Mapping;
+namespace Neo4j.Driver.Mapping;
 
 internal static class DefaultMapper
 {
