+++ Neo4j.Driver/Neo4j.Driver/Public/Mapping/DictAsRecord.cs
@@ -20,7 +20,7 @@ using System.Collections;
 using System.Collections.Generic;
 using System.Linq;
 
-namespace Neo4j.Driver.Preview.Mapping;
+namespace Neo4j.Driver.Mapping;
 
 internal sealed class DictAsRecord : IRecord
 {
