+++ Neo4j.Driver/Neo4j.Driver/Public/Mapping/EntityMappingInfo.cs
@@ -16,7 +16,7 @@
 using System.Linq;
 using System.Reflection;
 
-namespace Neo4j.Driver.Preview.Mapping;
+namespace Neo4j.Driver.Mapping;
 
 internal record EntityMappingInfo(
     string Path,
