+++ Neo4j.Driver/Neo4j.Driver/Public/Mapping/MappedListCreator.cs
@@ -17,7 +17,7 @@ using System;
 using System.Collections;
 using System.Collections.Generic;
 
-namespace Neo4j.Driver.Preview.Mapping;
+namespace Neo4j.Driver.Mapping;
 
 internal interface IMappedListCreator
 {
