+++ Neo4j.Driver/Neo4j.Driver/Public/Mapping/MappingExtensions.cs
@@ -17,7 +17,7 @@ using System;
 using System.Collections.Generic;
 using System.Reflection;
 
-namespace Neo4j.Driver.Preview.Mapping;
+namespace Neo4j.Driver.Mapping;
 
 internal static class MappingExtensions
 {
