+++ Neo4j.Driver/Neo4j.Driver/Public/Mapping/RecordExtensions.cs
@@ -15,7 +15,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-namespace Neo4j.Driver.Preview.Mapping;
+namespace Neo4j.Driver.Mapping;
 
 /// <summary>
 /// Contains extensions for accessing values simply from records and entities.
